Add Virtual-On mode
Implement Virtual-On mode that emulates the twin stick controller
by single arcade controller that has only one 1 stick and 4 buttons.

The arcade stick affects both left and right sticks, but if button 2
is pressed, right results in turning left, left does turning right,
up does jumping, and down does cancelling it. Button 1 is L shot,
button 4 is R shot, and button 3 is turbo for both L and R.

Not sure if this is the best assignments, but it works enough.
Suggestion is welcomed, but definitely using SS twin stick is the
best alternative way to play Virtual-On.