Bug fix and refactoring to support Arduino MEGA correctly
There is a bug that a wrong pin is assigned for the JVS sense
on Arduino MEGA. The bug was fixed in jvsio library, and this
commit contains the jvsio library roll, and iona.ino side fix
to avoid using the conflicted pin for the SENSE.

As clean-up, this patch move the pins array to the global space
and use the definition to scan button inputs.

Warning: this revision hasn't well verified.

See also https://github.com/toyoshim/iona/issues/10